requirements.txt
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-google-sso==9.0.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
httplib2==0.31.0
idna==3.11
loguru==0.7.3
oauthlib==3.3.1
pillow==12.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.5
django-environ==0.10.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.4
urllib3==2.6.2
